#57 Move a bit the code around, maybe is cleaner
